requirements.txt
# Core dependencies for NFL Play Retrieval Agent
# Python 3.9+

# Vector search
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# LLM
openai>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# BM25 for hybrid retrieval
rank-bm25>=0.2.2
